+++ Projeto 2-  com OO/Medicamento.h
@@ -11,7 +11,7 @@ class Medicamento : public Insumo {
 		std::string disponibilizacao;
 	public:
 		Medicamento();
-		Medicamento(std::string nome, int quantidade, int valorUnitario,
+        Medicamento(std::string nome, int quantidade, double valorUnitario,
                             std::string dtVencimento, std::string nomeFabricante, 
 							std::string dosagem, std::string administracao, std::string disponibilizacao); 
 		std::string getDisponibilizacao();
